- Updated LNL to 0.9.3.2 - Added channels. - Fixed some garbage creation. - Fixed problems with creating local connections in listen server. - Fixed problems with disconnecting and reconnecting. - Changed when the network packages are read from Update to FixedUpdate. - Added option for components to have their serialization sent by different methods and channels. - Added a priority queue (still not being used). - Fixed a problem with PostProcessor and playing while in the editor. - TinyNetGameManagerEditor now shows fields from child classes.

+++ Assets/TinyBirdNet/Example/Player/ExamplePawn.cs
@@ -22,7 +22,7 @@ public class ExamplePawn : TinyNetBehaviour {
 	byte netDir { get; set; }
 
 	[TinyNetSyncVar]
-	public short ownerPlayerControllerId { get; set; }
+	public byte ownerPlayerControllerId { get; set; }
 
 	public GameObject bulletPrefab;
